some refactoring such that the user file is not too bloated
